Inserted User Icon
Created and inserted icon which will be used to link users to the login
page.

The icon being used was found on a 3rd party website that provides
royalty free images IF credit is given to the creator. I did this by
inserting the provided link in the footer, just below the copyright
text. I may make a custom icon if able.

Also changed font size of the h1 font on the sitemap page from 50pt to
50px.